backend/requirements.txt
flask
faiss-cpu
sentence-transformers
